requirements.txt
fastapi
uvicorn
python-dotenv
openai
faiss-cpu
pypdf
pdf2image
pytesseract
Pillow
python-docx
python-pptx
pandas
markdown-it-py
orjson
llama-index
llama-index-embeddings-openai
llama-index-vector-stores-faiss
langchain
langchain-openai
langchain-community
